pyproject.toml
[project]
name = "libot2"
version = "0.1.0"
description = "libot2"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "blivedm @ git+https://github.com/xfgryujk/blivedm.git@master",
    "pillow>=10.0.0",
    "pilmoji>=2.0.4",
    "nonebot2[fastapi]>=2.5.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
]

[project.scripts]
libotctl = "src.libotctl:main"

[tool.nonebot]
plugin_dirs = []
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = ["src.plugins.libot"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
